refactor(analytics): extract analytics loading logic
